% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_algorithm.R
\name{compute_FDP_finite_est}
\alias{compute_FDP_finite_est}
\title{Return FDP_estimate without thresholding}
\usage{
compute_FDP_finite_est(Z, sl, sr)
}
\arguments{
\item{Z}{Numeric vector of z-values}

\item{sl}{Left-region threshold}

\item{sr}{Right-region threshold}
}
\value{
FDP_finite estimate
}
\description{
Return FDP_estimate without thresholding
}
\examples{
Z <- c(0.213, 1.652, 0.758, -1.149, -0.664)
compute_FDP_finite_est(Z, rep(0.15, 5), rep(0.85, 5))
## returns 1
}
