% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_run.R
\name{EM_run}
\alias{EM_run}
\title{Title}
\usage{
EM_run(
  Zs,
  is_masked,
  X,
  params_init,
  hyp_params,
  maxit = 200,
  tol = 1e-04,
  gating_option = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{params_init}{list of initial values for parameters (w0, w, beta0, beta, sigma2)}

\item{hyp_params}{list of fixed hyperparameter values (K, p, lambda, gamma)}

\item{gating_option}{If TRUE, uses Proximal Newton-type Method; else Proximal Newton.}
}
\value{
list of parameter estimates (w0, w, beta0, beta, sigma2)
}
\description{
Title
}
