% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_by}
\alias{scale_by}
\title{Scale vector}
\usage{
scale_by(Z, mean, sd)
}
\arguments{
\item{Z}{Numeric vector}

\item{mean}{Mean to scale with}

\item{sd}{Standard dev. to scale with}
}
\value{
Rescaled Z
}
\description{
Scale vector
}
