% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_column_subset}
\alias{scale_column_subset}
\title{Scale a matrix while leaving a subset of columns unchanged}
\usage{
scale_column_subset(X, exclude)
}
\arguments{
\item{X}{Matrix}

\item{exclude}{Indices of columns to exclude}
}
\value{
Xs, standardised only on columns -exclude
}
\description{
Scale a matrix while leaving a subset of columns unchanged
}
