% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_algorithm.R
\name{zap_v2}
\alias{zap_v2}
\title{ZAP algorithm using Regularised Gaussian Mixture-of-Expert Model}
\usage{
zap_v2(
  Z,
  X,
  K,
  lambda,
  gamma,
  alpha = 0.05,
  nfits = 50,
  masking_method = "basic",
  sl_thresh = 0.2,
  alpha_m = NA,
  nu = NA,
  lambda_m = NA,
  tol = 1e-04,
  maxit = 50,
  use_cpp = TRUE,
  use_proximal_newton = FALSE,
  EM_verbose = FALSE,
  zap_verbose = FALSE,
  seed = 1
)
}
\arguments{
\item{Z}{numeric vector of observed z-values to fit. TODO - assume scaled for unit variance, but not zero mean.}

\item{X}{n x p matrix-like object; n vectors of p predictors (standardised), EXCLUDING intercept col}

\item{alpha}{targeted FDR level.  Default to 0.05.}

\item{nfits}{maximum number of EM updates during the procedure... DEFAULT?}

\item{sl_thresh}{Hyperparameter for non-interactive threshold}

\item{maxit}{maximum number of iterations for every EM update}

\item{use_proximal_newton}{If TRUE, uses Proximal Newton Method; else Proximal Newton-type.}
}
\value{
Vector of indices for samples to reject
}
\description{
ZAP algorithm using Regularised Gaussian Mixture-of-Expert Model
}
