% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which.max.with_ties}
\alias{which.max.with_ties}
\title{Return all element indices attaining vector maxima}
\usage{
which.max.with_ties(x)
}
\arguments{
\item{x}{Numerical vector}
}
\value{
Vector of indices
}
\description{
Return all element indices attaining vector maxima
}
\examples{
which.max.with_ties(c(10,2,10,3))  # returns [1] 1 3
}
