% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_algorithm.R
\name{zap_v2}
\alias{zap_v2}
\title{ZAP algorithm using RGMOE z-value model}
\usage{
zap_v2(
  Z,
  X,
  alpha = 0.05,
  alpha_m = 0.2,
  K,
  lambda,
  gamma,
  maxit = 10,
  tol = 1e-04,
  nfits = 100,
  use_proximal_newton = FALSE,
  zap_verbose = TRUE,
  EM_verbose = FALSE
)
}
\arguments{
\item{Z}{numeric vector of observed z-values to fit}

\item{X}{dataframe of predictors (standardised), EXCLUDING intercept col}

\item{alpha}{targeted FDR level.  Default to 0.05.}

\item{alpha_m}{Hyperparameter for non-interactive threshold}

\item{maxit}{maximum number of iterations for every EM update}

\item{nfits}{maximum number of EM updates during the procedure... DEFAULT?}

\item{use_proximal_newton}{If TRUE, uses Proximal Newton Method; else Proximal Newton-type.}
}
\value{
Vector of indices for samples to reject
}
\description{
ZAP algorithm using RGMOE z-value model
}
