% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_algorithm.R
\name{zap_v2}
\alias{zap_v2}
\title{ZAP algorithm using Regularised Gaussian Mixture-of-Expert Model}
\usage{
zap_v2(
  Z,
  X,
  K,
  lambda,
  gamma,
  alpha = 0.05,
  alpha_m = 0.2,
  maxit = 50,
  masking_method = 1,
  tol = 1e-04,
  nfits = 50,
  use_proximal_newton = FALSE,
  zap_verbose = FALSE,
  EM_verbose = FALSE
)
}
\arguments{
\item{Z}{numeric vector of observed z-values to fit. TODO - assume scaled for unit variance, but not zero mean.}

\item{X}{n x p matrix-like object; n vectors of p predictors (standardised), EXCLUDING intercept col}

\item{alpha}{targeted FDR level.  Default to 0.05.}

\item{alpha_m}{Hyperparameter for non-interactive threshold}

\item{maxit}{maximum number of iterations for every EM update}

\item{nfits}{maximum number of EM updates during the procedure... DEFAULT?}

\item{use_proximal_newton}{If TRUE, uses Proximal Newton Method; else Proximal Newton-type.}
}
\value{
Vector of indices for samples to reject
}
\description{
ZAP algorithm using Regularised Gaussian Mixture-of-Expert Model
}
