% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_algorithm.R
\name{zap_v2}
\alias{zap_v2}
\title{ZAP algorithm using RGMOE z-value model}
\usage{
zap_v2(
  Xs,
  Z,
  alpha = 0.05,
  alpha_m = 0.2,
  hyp_params,
  params_init,
  EM_gating_option = F,
  zap_verbose = T,
  EM_verbose = F
)
}
\arguments{
\item{Xs}{dataframe of predictors (standardised), EXCLUDING intercept col}

\item{Z}{numeric vector of observed z-values to fit}

\item{alpha}{FDR control level}

\item{alpha_m}{Hyperparameter for non-interactive threshold}
}
\value{
Vector of indices for samples to reject
}
\description{
ZAP algorithm using RGMOE z-value model
}
